requirements.txt
flask==3.0.3
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3